[#31] Accommodation에 User 권한(Role) 추가 (#35)
* chore: naver-checkstyle-rules 적용

* style: check style에 따른 indent 수정

* refactor: controller json api 반환 형식 변경

* refactor: VerificationUtils 생성

* feat: User 필드 추가

* feat: User 권한 추가 (@LoginCheck, AuthUser, @RequireLoginUser)

* test: User 권한 추가 후 기존 테스트 코드 수정

* refactor: Util 클래스 private 생성자 추가

* refactor: service layer 파라미터 domain -> dto로 변경

* style: owner로 필드명 변경

* refactor: requestDto에 owner 필드 추가

* test: service 계층 파라미터 dto로 변경 후 테스트

* refactor: User 객체 참조를 id 참조로 변경

+++ src/main/java/com/project/jagoga/accommodation/application/AccommodationService.java
@@ -1,11 +1,14 @@
 package com.project.jagoga.accommodation.application;
 
 import com.project.jagoga.accommodation.presentation.dto.AccommodationRequestDto;
-import com.project.jagoga.accommodation.presentation.dto.AccommodationResponseDto;
 import com.project.jagoga.exception.accommodation.DuplicatedAccommodationException;
 import com.project.jagoga.exception.accommodation.NotExistAccommodationException;
 import com.project.jagoga.accommodation.domain.Accommodation;
 import com.project.jagoga.accommodation.domain.AccommodationRepository;
+import com.project.jagoga.user.domain.AuthUser;
+import com.project.jagoga.user.domain.User;
+import com.project.jagoga.user.domain.UserRepository;
+import com.project.jagoga.utils.VerificationUtils;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -16,42 +19,50 @@ import java.util.List;
 public class AccommodationService {
 
     private final AccommodationRepository accommodationRepository;
+    private final UserRepository userRepository;
 
-    public Long saveAccommodation(Accommodation accommodation) {
+    public Accommodation saveAccommodation(AccommodationRequestDto accommodationRequestDto, AuthUser loginUser) {
+        User owner = userRepository.findById(loginUser.getId()).get();
+        VerificationUtils.verifyPermission(loginUser, owner.getId());
+        Accommodation accommodation = accommodationRequestDto.toEntity(owner);
         validateDuplicatedAccommodation(accommodation);
-        return accommodationRepository.save(accommodation).getAccommodationId();
+        return accommodationRepository.save(accommodation);
     }
 
-    public AccommodationResponseDto updateAccommodation(
+    public Accommodation updateAccommodation(
         long accommodationId,
-        AccommodationRequestDto accommodationRequestDto
+        AccommodationRequestDto accommodationRequestDto,
+        AuthUser loginUser
     ) {
         Accommodation accommodation = accommodationRepository.findById(accommodationId)
-                .orElseThrow(NotExistAccommodationException::new);
-
+            .orElseThrow(NotExistAccommodationException::new);
+        Long ownerId = accommodation.getOwnerId();
+        VerificationUtils.verifyPermission(loginUser, ownerId);
         Accommodation updatedAccommodation = accommodation.update(
-                accommodationRequestDto.getAccommodationName(),
-                accommodationRequestDto.getPhoneNumber(),
-                accommodationRequestDto.getCity(),
-                accommodationRequestDto.getAccommodationType(),
-                accommodationRequestDto.getAccommodationName(),
-                accommodationRequestDto.getInformation());
-        return AccommodationResponseDto.of(accommodationRepository.update(updatedAccommodation));
+            accommodationRequestDto.getAccommodationName(),
+            accommodationRequestDto.getPhoneNumber(),
+            accommodationRequestDto.getCity(),
+            accommodationRequestDto.getAccommodationType(),
+            accommodationRequestDto.getAccommodationName(),
+            accommodationRequestDto.getInformation());
+        return accommodationRepository.update(updatedAccommodation);
     }
 
-    public Long deleteAccommodation(long accommodationId) {
+    public Long deleteAccommodation(long accommodationId, AuthUser loginUser) {
+        Long ownerId = accommodationRepository.findById(accommodationId).get().getOwnerId();
+        VerificationUtils.verifyPermission(loginUser, ownerId);
         accommodationRepository.findById(accommodationId)
-                .ifPresentOrElse(
-                    a -> accommodationRepository.delete(accommodationId),
-                    () -> {
-                        throw new NotExistAccommodationException();
-                    });
+            .ifPresentOrElse(
+                a -> accommodationRepository.delete(accommodationId),
+                () -> {
+                    throw new NotExistAccommodationException();
+                });
         return accommodationId;
     }
 
     public Accommodation getAccommodation(long accommodationId) {
         return accommodationRepository.findById(accommodationId)
-                .orElseThrow(NotExistAccommodationException::new);
+            .orElseThrow(NotExistAccommodationException::new);
     }
 
     public List<Accommodation> getAccommodationAllList() {
@@ -64,8 +75,8 @@ public class AccommodationService {
 
     private void validateDuplicatedAccommodation(Accommodation accommodation) {
         accommodationRepository.findByName(accommodation.getAccommodationName())
-                .ifPresent(a -> {
-                    throw new DuplicatedAccommodationException();
-                });
+            .ifPresent(a -> {
+                throw new DuplicatedAccommodationException();
+            });
     }
 }
